refactor(mobile): device 핸들러 도메인별 분리
